Dificultad añadida a las rutinas 1,2,3

+++ src/main/java/com/gimnsio/libreta/DTO/serie/SerieForWorkoutDTO.java
@@ -9,5 +9,5 @@ public class SerieForWorkoutDTO {
     private Long id;
     private String exercise;
     private int reps;
-    private int kg;
+    private double kg;
 }
